Backend/requirements.txt
django
djangorestframework
djoser
django-cors-headers
djangorestframework-simplejwt